requirements.txt
bs4
markdown
streamlit
requests
openai
pdfkit
python-pptx
python-dotenv
google-genai
googlesearch-python